Moving files -- prepare for shelley integration
