+++ app/frontend/wallet/cardano-wallet.ts
@@ -4,15 +4,15 @@ import {base58} from 'cardano-crypto.js'
 import debugLog from '../helpers/debugLog'
 import {generateMnemonic, validateMnemonic} from './mnemonic'
 
-import {TxInputFromUtxo, TxOutput, TxAux} from './byron-transaction'
+import {TxInputFromUtxo, TxOutput, TxAux} from './byron/byron-transaction'
 
 import AddressManager from './address-manager'
-import {ByronAddressProvider} from './byron-address-provider'
+import {ByronAddressProvider} from './byron/byron-address-provider'
 import BlockchainExplorer from './blockchain-explorer'
 import PseudoRandom from './helpers/PseudoRandom'
 import {ADA_DONATION_ADDRESS, MAX_INT32, TX_WITNESS_SIZE_BYTES} from './constants'
 import shuffleArray from './helpers/shuffleArray'
-import CborIndefiniteLengthArray from './helpers/CborIndefiniteLengthArray'
+import CborIndefiniteLengthArray from './byron/helpers/CborIndefiniteLengthArray'
 import NamedError from '../helpers/NamedError'
 import {roundWholeAdas} from '../helpers/adaConverters'
 import {Lovelace} from '../state'
