+++ app/frontend/wallet/helpers/hd-node.ts
