+++ app/frontend/wallet/helpers/indexIsHardened.ts
