+++ app/tests/src/address-manager.js
@@ -1,14 +1,14 @@
 import assert from 'assert'
 
-import derivationSchemes from '../../frontend/wallet/crypto-providers/derivation-schemes'
-import CardanoWalletSecretCryptoProvider from '../../frontend/wallet/crypto-providers/cardano-wallet-secret-crypto-provider'
+import derivationSchemes from '../../frontend/wallet/helpers/derivation-schemes'
+import CardanoWalletSecretCryptoProvider from '../../frontend/wallet/byron/cardano-wallet-secret-crypto-provider'
 import AddressManager from '../../frontend/wallet/address-manager'
 import mnemonicToWalletSecretDef from '../../frontend/wallet/helpers/mnemonicToWalletSecretDef'
 import cryptoProviderSettings from './common/crypto-provider-settings'
 import BlockchainExplorer from '../../frontend/wallet/blockchain-explorer'
 
 import mockNetwork from './common/mock'
-import {ByronAddressProvider} from '../../frontend/wallet/byron-address-provider'
+import {ByronAddressProvider} from '../../frontend/wallet/byron/byron-address-provider'
 
 const mockConfig = {
   ADALITE_BLOCKCHAIN_EXPLORER_URL: 'https://explorer.adalite.io',
