+++ app/tests/src/cardano-wallet-secret-crypto-provider.js
@@ -2,9 +2,9 @@ import assert from 'assert'
 import cbor from 'borc'
 
 import {HARDENED_THRESHOLD} from '../../frontend/wallet/constants'
-import derivationSchemes from '../../frontend/wallet/crypto-providers/derivation-schemes'
-import CardanoWalletSecretCryptoProvider from '../../frontend/wallet/crypto-providers/cardano-wallet-secret-crypto-provider'
-import {TxInputFromUtxo, TxOutput, TxAux} from '../../frontend/wallet/byron-transaction'
+import derivationSchemes from '../../frontend/wallet/helpers/derivation-schemes'
+import CardanoWalletSecretCryptoProvider from '../../frontend/wallet/byron/cardano-wallet-secret-crypto-provider'
+import {TxInputFromUtxo, TxOutput, TxAux} from '../../frontend/wallet/byron/byron-transaction'
 import mnemonicToWalletSecretDef from '../../frontend/wallet/helpers/mnemonicToWalletSecretDef'
 import cryptoProviderSettings from './common/crypto-provider-settings'
